restructure folders to make files smaller
