+++ hello_books/api/auth_views.py
@@ -6,28 +6,26 @@ from flask_jwt_extended import (
     create_access_token, get_raw_jwt
 )
 from werkzeug.security import generate_password_hash, check_password_hash
-from hello_books import app, jwt
-from hello_books.models import HelloBooks,User
-
-blacklist = set()
-
-@jwt.token_in_blacklist_loader
-def check_if_token_in_blacklist(decrypted_token):
-    '''check if token is blacklisted'''
-    jti = decrypted_token['jti']
-    return jti in blacklist
+from hello_books import create_app
+from hello_books.models.validate_model import HelloBooks
+from hello_books.models.book_model import Books
+from hello_books.models.user_model import User
 
 auth = Blueprint('auth', __name__)
-hello_books = HelloBooks()
 
-@app.route('/', methods=['GET'])
+
+
+
+
+
+@auth.route('/', methods=['GET'])
 def home():
     '''Home page containing link to documentation'''
     return jsonify(
         {'Hello Books API': "Click here to see documentation -> https://hellobooks8.docs.apiary.io/"})
 
 
-@app.route('/api/v1/auth/reset-password', methods=['POST'])
+@auth.route('/api/v1/auth/reset-password', methods=['POST'])
 def reset_password():
     '''Function to reseta user password'''
     email = request.json.get('email').strip()
@@ -40,7 +38,7 @@ def reset_password():
     }), 404
     
 
-@app.route('/api/v1/auth/register', methods=['POST'])
+@auth.route('/api/v1/auth/register', methods=['POST'])
 def register():
     '''Fuction to register a new user'''
     try:
@@ -54,7 +52,7 @@ def register():
         if User().check_email_exists(data['email']) == True:
             return jsonify({'message': 'Email Exists'})
         else:
-            if hello_books.user_data_validation(data) == True:
+            if HelloBooks().user_data_validation(data) == True:
                 new_user = User(
                     username = data['name'],
                     email=data['email'],
@@ -69,26 +67,22 @@ def register():
         return jsonify({"message": "An error occured. Please try again."}), 500
 
 
-@app.route('/api/v1/auth/login', methods=['POST'])
+@auth.route('/api/v1/auth/login', methods=['POST'])
 def login():
     '''Function for logging in'''
-    try:
-        sent_data = request.get_json(force=True)
-        raw_data = {
-            'email': sent_data['email'],
-            'password': sent_data['password']
-        }
-        data = {k.strip(): v.strip() for k, v in raw_data.items()}
-        return User().user_login(
-            mail=data['email'],
-            password=data['password']
-        )
-    except BaseException:
-        return jsonify(
-            {'message': 'Please enter your email and password correctly'})
+    sent_data = request.get_json(force=True)
+    raw_data = {
+        'email': sent_data['email'],
+        'password': sent_data['password']
+    }
+    data = {k.strip(): v.strip() for k, v in raw_data.items()}
+    return User().user_login(
+        mail=data['email'],
+        password=data['password']
+    )
 
 
-@app.route('/api/v1/auth/logout', methods=['GET', 'POST'])
+@auth.route('/api/v1/auth/logout', methods=['GET', 'POST'])
 @jwt_required
 def logout():
     '''Function for logout'''
@@ -97,7 +91,7 @@ def logout():
     return jsonify({'message': 'You are now logged out'}), 200
 
 
-@app.route('/api/v1/auth/change-password', methods=['PUT'])
+@auth.route('/api/v1/auth/change-password', methods=['PUT'])
 @jwt_required
 def change_password():
     '''Function for changing user password'''
@@ -113,7 +107,7 @@ def change_password():
         return jsonify({"message": "An error occured. Please try again."})
 
 
-@app.route('/api/v1/auth/users', methods=['GET'])
+@auth.route('/api/v1/auth/users', methods=['GET'])
 @jwt_required
 def view_users():
     '''Function for viewing all books'''
@@ -128,7 +122,7 @@ def view_users():
     return User().view_users(page=page, per_page=results), 200
 
 
-@app.route('/api/v1/auth/make-admin', methods=['PUT'])
+@auth.route('/api/v1/auth/make-admin', methods=['PUT'])
 @jwt_required
 def make_admin():
     '''Function for changing a user to an admin'''
@@ -144,7 +138,7 @@ def make_admin():
         return jsonify({"message": "An error occured. Please try again."})
 
 
-@app.route('/api/v1/auth/authorize', methods=['PUT'])
+@auth.route('/api/v1/auth/authorize', methods=['PUT'])
 @jwt_required
 def authorize():
     '''Function for changing a user to an admin'''
