+++ hello_books/models/validate_model.py
@@ -1,0 +1,134 @@
+from cerberus import Validator
+import datetime
+
+
+class HelloBooks(object):
+    '''Class for data validation'''
+
+    def user_data_validation(self, dict_data):
+        '''user data validation method'''
+        schema = {
+            'name': {
+                'type': 'string',
+                'required': True,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'empty': False,
+                'maxlength': 20,
+                'minlength': 4},
+            'email': {
+                'type': 'string',
+                'regex': '^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$'},
+            'password': {
+                'type': 'string',
+                'required': True,
+                'maxlength': 16,
+                'minlength': 6}}
+        v = Validator(schema)
+        v.allow_unknown = True
+        return v.validate(dict_data)
+
+    def password_validation(self, dict_data):
+        '''Password validation method'''
+        schema = {
+            'password': {
+                'type': 'string',
+                'required': True,
+                'maxlength': 16,
+                'minlength': 6}}
+        v = Validator(schema)
+        v.allow_unknown = True
+        return v.validate(dict_data)
+
+    def add_book_validation(self, dict_data):
+        '''book data validation function'''
+        schema = {
+            'title': {
+                'type': 'string',
+                'required': True,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'empty': True,
+                'maxlength': 25,
+                'minlength': 4},
+            'author': {
+                'type': 'string',
+                'required': True,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'empty': True,
+                'maxlength': 25,
+                'minlength': 4},
+            'genre': {
+                'type': 'string',
+                'required': True,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'empty': True,
+                'maxlength': 20,
+                'minlength': 4},
+            'isbn': {
+                'type': 'string',
+                'required': True,
+                'regex': '^[0-9]+$',
+                'empty': False,
+                'maxlength': 13,
+                'minlength': 13},
+            'description': {
+                'type': 'string',
+                'required': True,
+                'maxlength': 200,
+                'minlength': 4}}
+        v = Validator(schema)
+        v.allow_unknown = True
+        return v.validate(dict_data)
+
+    def edit_book_validation(self, dict_data):
+        '''book data validation function using CERBERUS'''
+        schema = {
+            'title': {
+                'type': 'string',
+                'required': False,
+                'empty': True,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'maxlength': 25,
+                'minlength': 4},
+            'author': {
+                'type': 'string',
+                'required': False,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'empty': True,
+                'maxlength': 25,
+                'minlength': 4},
+            'genre': {
+                'type': 'string',
+                'required': False,
+                'regex': '^[a-zA-Z0-9 ]+$',
+                'empty': True,
+                'maxlength': 20,
+                'minlength': 4},
+            'copies': {
+                'type': 'string',
+                'required': False,
+                'regex': '^[0-9]+$',
+                'empty': False,
+                'maxlength': 1,
+                'minlength': 3},
+            'isbn': {
+                'type': 'string',
+                'required': False,
+                'regex': '^[0-9]+$',
+                'empty': False,
+                'maxlength': 13,
+                'minlength': 13},
+            'description': {
+                'type': 'string',
+                'required': False,
+                'maxlength': 200,
+                'minlength': 4}}
+        v = Validator(schema)
+        v.allow_unknown = True
+        return v.validate(dict_data)
+
+    def date_validate(self, date_text):
+        try:
+            datetime.datetime.strptime(date_text, '%d/%m/%Y')
+            return True
+        except BaseException:
+            return False
