+++ requirements.txt
@@ -19,6 +19,7 @@ Flask-Migrate==2.1.1
 Flask-RESTful==0.3.6
 Flask-Script==2.0.6
 Flask-SQLAlchemy==2.3.2
+Flask-Testing==0.7.1
 idna==2.6
 isort==4.3.4
 itsdangerous==0.24
