+++ run.py
@@ -1,2 +1,7 @@
-from hello_books import app
-app.run(debug=True)
+from hello_books import create_app
+
+config_name = 'development'
+app = create_app(config_name)
+
+if __name__ == '__main__':
+    app.run()
